Added editor selection conditions matching chains
To match the route pattern adds the ability to set the JavaScript
condition to select the desired route.

The function should return a Boolean value.

It is assumed that the condition must be unique. If you create a template 2,
that, when the condition returns true, then picked up the first
comer pattern.

Example.
'Return document.properties [ "mirm: targetQ4"] <3000;' - The template will be
applied to the document, which mirm field: targetQ4 less than 3000.

--HG--
branch : test